C#/Lista1/AbstractFactory/Report.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using AbstractFactory.json;
using AbstractFactory.xml;

namespace AbstractFactory
{
    public class Report {

	private String reportContent;
	private ReportBody body;
	private ReportFooter footer;
	private ReportHeader header;
	private String reportType;
	

	
	
	/**
	 * @param string
	 */
    public Report(AbstractReportElementsFactory factory)
    {
        this.setBody(factory.createReportBody());

        this.setFooter(factory.createReportFooter());

        this.setHeader(factory.createReportHeader());
    }


	public void setBody(ReportBody body) {
		this.body = body;

	}

	
	public void setFooter(ReportFooter footer) {
		this.footer = footer;

	}

	
	public void setHeader(ReportHeader header) {
		this.header = header;
	}

	public void setReportContent(String reportContent) {
		this.reportContent = reportContent;
	}


	public String getReportContent() {
		return reportContent;
	}


	public ReportBody getBody() {
		return body;
	}


	public ReportFooter getFooter() {
		return footer;
	}


	public ReportHeader getHeader() {
		return header;
	}

	
}
}
